エクスポート処理追加

+++ Rialto/Models/Service/ExportService.cs
@@ -12,12 +12,15 @@ using System.Text;
 using System.Threading.Tasks;
 using LanguageExt;
 using static LanguageExt.Prelude;
+using NLog;
 
 namespace Rialto.Models.Service
 {
     public class ExportService
     {
-        public Task<IEnumerable<Try<string>>> Export(string exportDir, long tagId)
+        private static readonly Logger logger = LogManager.GetLogger("fileLogger");
+
+        public Task<List<Try<string>>> Export(string exportDir, long tagId)
         {
             if (Directory.Exists(exportDir))
             {
@@ -27,13 +30,22 @@ namespace Rialto.Models.Service
                         return list.Select(item =>
                         (from registerImage in item.Item1
                          from repository in item.Item2
-                         select (Path.Combine(repository.Path, registerImage.FilePath), registerImage.FileName))
-                             .Fold(Try<string>(new Exception("Option None")), (acc, x) => {
+                         select (Path.Combine(repository.Path, registerImage.FilePath), $"{registerImage.FileName}.{registerImage.FileExtension}"))
+                             .Fold(Try<string>(new Exception("Option None")), (acc, x) =>
+                             {
                                  (string sourcePath, string fileName) = x;
                                  try
                                  {
                                      var destFilePath = Path.Combine(exportDir, fileName);
-                                     File.Copy(sourcePath, destFilePath);
+                                     if (File.Exists(destFilePath))
+                                     {
+                                         // スキップ
+                                         logger.Debug($"Export Skip = {destFilePath}");
+                                     } else
+                                     {
+                                         File.Copy(sourcePath, destFilePath);
+                                     }
+                                     
                                      return Try(destFilePath);
                                  }
                                  catch (Exception e)
@@ -41,12 +53,13 @@ namespace Rialto.Models.Service
                                      return Try<string>(e);
                                  }
                              })
-                        );
+                        ).ToList();
                     });
                 });
             } else
             {
-                return Task.FromResult(Enumerable.Empty<Try<string>>());
+                
+                return Task.FromResult(new System.Collections.Generic.List<Try<string>>());
             }
         }
